+++ Gridlet.h
@@ -30,7 +30,7 @@ namespace exa {
     inline __both__
     box3f getBounds() const
     {
-      const float cellWidth = 1<<level;
+      const float cellWidth = (float)(1<<level);
       return box3f((vec3f(lower) + 0.5f) * cellWidth,
                    (vec3f(lower) + vec3f(dims) + 0.5f) * cellWidth);
     }
